+++ src/toParts.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
 import { toParts } from './toParts.js';
-import { weeks } from './conversion.js';
 
 describe('toParts()', () => {
 	it('returns the correct parts for a positive duration', () => {
@@ -10,8 +9,10 @@ describe('toParts()', () => {
 			days: 1,
 			hours: 20,
 			isNegative: false,
+			microseconds: 0,
 			milliseconds: 500,
 			minutes: 56,
+			nanoseconds: 0,
 			seconds: 46,
 			weeks: 1,
 		});
@@ -24,8 +25,42 @@ describe('toParts()', () => {
 			days: 1,
 			hours: 20,
 			isNegative: true,
+			microseconds: 0,
 			milliseconds: 500,
 			minutes: 56,
+			nanoseconds: 0,
+			seconds: 46,
+			weeks: 1,
+		});
+	});
+
+	it('returns the correct parts for a positive duration with decimal places', () => {
+		const result = toParts(766_606_500.003002);
+
+		expect(result).toEqual({
+			days: 1,
+			hours: 20,
+			isNegative: false,
+			microseconds: 3,
+			milliseconds: 500,
+			minutes: 56,
+			nanoseconds: 2,
+			seconds: 46,
+			weeks: 1,
+		});
+	});
+
+	it('returns the correct parts for a negative duration with decimal places', () => {
+		const result = toParts(-766_606_500.003002);
+
+		expect(result).toEqual({
+			days: 1,
+			hours: 20,
+			isNegative: true,
+			microseconds: 3,
+			milliseconds: 500,
+			minutes: 56,
+			nanoseconds: 2,
 			seconds: 46,
 			weeks: 1,
 		});